removing old files
